Move Search component to separate package
