+++ src/js/components/header/Header.jsx
@@ -1,7 +1,7 @@
 import React from 'react/addons';
 
 import Logo from './Logo';
-import Search from './Search';
+import Search from './search/Search';
 import UserMenu from './UserMenu';
